Challenges id is fetched.
