+++ src/components/ChallengeList/ChallengeList.js
@@ -5,13 +5,13 @@ import { Route, Switch, useRouteMatch } from "react-router";
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
 import { Typography } from "@mui/material";
-import Toolbar from '@mui/material/Toolbar';
-import CssBaseline from '@mui/material/CssBaseline';
-import useScrollTrigger from '@mui/material/useScrollTrigger';
-import Box from '@mui/material/Box';
-import Fab from '@mui/material/Fab';
-import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
-import Zoom from '@mui/material/Zoom';
+import Toolbar from "@mui/material/Toolbar";
+import CssBaseline from "@mui/material/CssBaseline";
+import useScrollTrigger from "@mui/material/useScrollTrigger";
+import Box from "@mui/material/Box";
+import Fab from "@mui/material/Fab";
+import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
+import Zoom from "@mui/material/Zoom";
 
 // styling
 import { useStyles } from "./styles";
@@ -19,7 +19,7 @@ import { useStyles } from "./styles";
 // custom components
 import SensexChart from "./SensexChart";
 import ChallengeFilter from "./ChallengeFilter";
-import ChallengeCard from "./ChallengeCard";
+import ChallengeCard from "./ChallengeCard/ChallengeCard";
 import Challenge from "../Challenge/Challenge";
 import Header from "../Header/Header";
 
@@ -59,7 +59,6 @@ function ScrollTop(props) {
 }
 
 export default function ChallengeList(props) {
-
   const classes = useStyles();
   const [challenges, setChallenges] = useState([]);
   const { path } = useRouteMatch();
@@ -68,7 +67,9 @@ export default function ChallengeList(props) {
     const fetchData = async () => {
       try {
         // Fetch data from REST API
-        const response = await fetch("http://localhost:8080/api/challenge/filter/STARTING_SOON");
+        const response = await fetch(
+          "http://localhost:8080/api/challenge/filter/STARTING_SOON"
+        );
         if (response.status === 200) {
           // Extract json
           const rawData = await response.json();
@@ -101,17 +102,25 @@ export default function ChallengeList(props) {
                     <SensexChart />
                   </Grid>
                   <Grid item xs={12} elevation={10}>
-
-                    <Typography variant="h4" className={classes.challengeListTitle}>Pick your challenge</Typography>
-                    <Grid container direction="row" spacing={3} className={classes.challengeList}>
-                      {
-                        challenges.map((item, index) => {
-                          return (
-                            <Grid item xs="12" md="12" lg="6" key={index}>
-                              <ChallengeCard challenge={item} />
-                            </Grid>
-                          );
-                        })}
+                    <Typography
+                      variant="h4"
+                      className={classes.challengeListTitle}
+                    >
+                      Pick your challenge
+                    </Typography>
+                    <Grid
+                      container
+                      direction="row"
+                      spacing={3}
+                      className={classes.challengeList}
+                    >
+                      {challenges.map((item, index) => {
+                        return (
+                          <Grid item xs="12" md="12" lg="6" key={index}>
+                            <ChallengeCard challenge={item} />
+                          </Grid>
+                        );
+                      })}
                     </Grid>
                   </Grid>
                 </Grid>
@@ -124,7 +133,7 @@ export default function ChallengeList(props) {
             </Fab>
           </ScrollTop>
         </Route>
-        <Route path={`${path}/challenge`}>
+        <Route path={`${path}/challenge:id`}>
           <Challenge />
         </Route>
       </Switch>
